requirements.txt
playwright
pytest
pytest-bdd
allure-pytest
faker
python-dotenv
pytest-html